fix: Glossary terms in .md url
