+++ lib/markdown-component-renderers.js
@@ -1,0 +1,48 @@
+const { CATEGORIES, TERMS } = require("./glossary-data.js");
+
+function replaceComponentsWithMarkdown(fileContent) {
+  return fileContent.replace(
+    /<Glossary\s*\/>/g,
+    () => `\n${renderGlossary()}\n`,
+  );
+}
+
+function renderGlossary() {
+  const sortedTerms = TERMS.toSorted((a, b) => a.term.localeCompare(b.term));
+  let currentLetter = "";
+  const lines = [];
+
+  for (const term of sortedTerms) {
+    const letter = term.term[0].toUpperCase();
+    if (letter !== currentLetter) {
+      currentLetter = letter;
+      lines.push(`## ${letter}`, "");
+    }
+
+    lines.push(`### ${term.term} [#${term.id}]`, "");
+    if (term.synonyms?.length) {
+      lines.push(`Also known as: ${term.synonyms.join(", ")}`, "");
+    }
+    lines.push(term.definition, "");
+    if (term.categories?.length) {
+      lines.push(
+        `Categories: ${term.categories
+          .map((category) => CATEGORIES[category]?.label ?? category)
+          .join(", ")}`,
+        "",
+      );
+    }
+    if (term.relatedTerms?.length) {
+      lines.push(`Related: ${term.relatedTerms.join(", ")}`, "");
+    }
+    if (term.link) {
+      lines.push(`[Learn more](${term.link})`, "");
+    }
+  }
+
+  return lines.join("\n");
+}
+
+module.exports = {
+  replaceComponentsWithMarkdown,
+};
